+++ packages/vitdoc-runtime/src/index.tsx
@@ -1,29 +1,14 @@
-import React, { Suspense, lazy } from "react";
+import React from "react";
 import { createRoot } from "react-dom/client";
 import { HashRouter, Route, Routes } from "react-router-dom";
-
-const Pure = lazy(() => import("./pure"));
-const Main = lazy(() => import("./main"));
+import Main from "./main";
+import Pure from "./pure";
 
 createRoot(document.querySelector("#component-root")!).render(
   <HashRouter>
     <Routes>
-      <Route
-        path="~/*"
-        element={
-          <Suspense fallback="Loading...">
-            <Pure />
-          </Suspense>
-        }
-      />
-      <Route
-        path="*"
-        element={
-          <Suspense fallback="Loading...">
-            <Main />
-          </Suspense>
-        }
-      />
+      <Route path="~/*" element={<Pure />} />
+      <Route path="*" element={<Main />} />
     </Routes>
   </HashRouter>
 );