+++ packages/vitdoc-runtime/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @vitdoc/runtime
 
+## 1.4.6-beta.6
+
+### Patch Changes
+
+- upgrade the performance
+- Updated dependencies
+  - @vitdoc/ui@1.3.5-beta.6
+
 ## 1.4.6-beta.5
 
 ### Patch Changes