+++ packages/vitdoc-ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vitdoc/ui
 
+## 1.3.5-beta.6
+
+### Patch Changes
+
+- upgrade the performance
+
 ## 1.3.5-beta.5
 
 ### Patch Changes